+++ All.ui/pages/index/config.js
@@ -1,0 +1,22 @@
+import { SVG } from '../../utils/config-svg'
+import { changeColor } from '../../utils/util'
+export const list = [
+  {
+    title: 'UI组件',
+    list: [
+      {
+        svg: SVG.nprogress,
+        name: '顶部加载'
+      },
+      {
+        svg: SVG.calendar,
+        name: '日历'
+      },
+      {
+        svg: SVG.colorPicker,
+        name: '颜色选择器'
+      }
+    ]
+  }
+
+]