+++ All.ui/pages/index/index.js
@@ -1,12 +1,11 @@
 // index.js
 // 获取应用实例
 const app = getApp()
-
+import { list } from './config'
 Page({
   data: {
-
+    list: list
   },
   onLoad() {
-    console.log('??')
   },
 })