+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v0.1
 
+### 2021/02/20
+- init theme config
+
 ### 2021/02/19
 - init All.ui